+++ example/android/app/src/main/java/flutter/plugins/flutter_vibrate_example/MainActivity.java
@@ -1,0 +1,6 @@
+package flutter.plugins.flutter_vibrate_example;
+
+import io.flutter.embedding.android.FlutterActivity;
+
+public class MainActivity extends FlutterActivity {
+}